requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
discord==1.0.1
discord.py==1.4.1
future==0.18.2
geojson==2.5.0
idna==2.10
lxml==4.5.2
multidict==4.7.6
oauthlib==3.1.0
pkg-resources==0.0.0
praw==7.1.0
prawcore==1.5.0
pyowm==3.0.0
PySocks==1.7.1
python-dotenv==0.14.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
tweepy==3.9.0
update-checker==0.18.0
urllib3==1.25.10
websocket-client==0.57.0
yarl==1.5.1
